ajout des fonction de connexion et inscription des professionel

+++ src/main/java/com/example/goodservice/repo/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserRegister, Long> {
+public interface UserRepository extends JpaRepository<UserRegister, Integer> {
     public UserRegister findByEmail(String email);
     public UserRegister findUserByEmailAndMdp(String email, String mdp);
 }
